Move serialization directory out of dags folder
(cherry picked from commit 8a07aee3e5cf133c45ee4ae26aad6104c84502ab)
