+++ airflow/serialization/serialized_dag.py
@@ -24,8 +24,8 @@ except ImportError:
     from funcsigs import signature  # type: ignore
 from typing import cast
 
-from airflow.dag.serialization.json_schema import load_dag_schema
-from airflow.dag.serialization.serialization import Serialization
+from airflow.serialization.json_schema import load_dag_schema
+from airflow.serialization.serialization import Serialization
 from airflow.models import DAG
 
 
@@ -92,7 +92,7 @@ class SerializedDAG(DAG, Serialization):
         # type: (dict) -> 'SerializedDAG'
         """Deserializes a DAG from a JSON object.
         """
-        from airflow.dag.serialization import SerializedBaseOperator
+        from airflow.serialization import SerializedBaseOperator
 
         dag = SerializedDAG(dag_id=encoded_dag['_dag_id'])
