+++ tests/models/test_serialized_dag.py
@@ -22,7 +22,7 @@
 import unittest
 
 from airflow import example_dags as example_dags_module
-from airflow.dag.serialization.serialized_dag import SerializedDAG
+from airflow.serialization.serialized_dag import SerializedDAG
 from airflow.models import DagBag
 from airflow.models import SerializedDagModel as SDM
 from airflow.utils import db
